. r: Update Basic Example to include Audio Player
